Major bug fixes, now it works

+++ lib/api/models/query.model.dart
@@ -5,14 +5,16 @@ import 'package:diskursv2/utils/types.dart';
 @immutable
 class QueryModel extends Equatable {
   final String? term;
-  final double? similarity;
+  final int? similarity;
   final double? frequency;
 
   const QueryModel._(this.term, this.similarity, this.frequency);
   factory QueryModel.fromJson(JSON json) => QueryModel._(
         json['term'],
-        json['freq'] != null ? (json['freq'] as double) * 100 : json['prob'],
-        json['avg'],
+        json['freq'] != null
+            ? ((json['freq'] as double) * 100).toInt()
+            : json['prob'] as int,
+        json['avg'] as double,
       );
 
   @override
